pyproject.toml
[project]
name = "mytaskly-mcp"
version = "0.1.0"
description = "MCP Server for MyTaskly with OAuth 2.1 authentication"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.5.0",
    "httpx>=0.27.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
